PDFBOX-4449: remove uneeded cast; apply Math.abs() to scale values
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1852534 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/rendering/PDFRenderer.java
@@ -441,7 +441,7 @@ public class PDFRenderer
                     break;
             }
             graphics.translate(translateX, translateY);
-            graphics.rotate((float) Math.toRadians(rotationAngle));
+            graphics.rotate(Math.toRadians(rotationAngle));
         }
     }
